feat(routes): Configure protected routes
To ensure that only Recursers have access to the contents of the site
(outside of home and overview), I've implemented protected routes.

The change required a decent overhaul of the src/ dir but I think it
makes sense for organizational purposes. The user should be redirected
to their target page after the redirect to /login (and subsequent
authentication through OAuth2).

You can find the link used to actually get these routes working here: https://stackoverflow.com/questions/59422159/redirecting-a-user-to-the-page-they-requested-after-successful-authentication-wi/59423442#59423442