refactor(misc): Separate project by server/client
To create a better split between the Flask and SQLite backend and then
React frontend, I've decided to modularize through the usage of client
and server directories.

The actual project content has not changed in any way but this allow for
a clear distinction between the two parts of the application.

The root now looks like:
  - client/
  - server/
  - .git
  - .gitignore