feat(UI): Start Home page implementation
Set the background to a stock RC image and created components (WelcomeCard
and WelcomeAccordion) to start the homepage implementation.

Still requires some CSS updating and overall styling fixes but it
contains the bulk of what will make up the finalized page.

Looking for a way to incorporate the RC logo as well as fade in the
primary card upon rendering of the page.