feat(UI): Implement carousel in Home
Finished most of Home's implementation, including an RC background and
links to the GitHub and an interal carousel. The carousel includes
explanations for each page as well as details about how to contribute.

Moving forward, I need to figure out fonts and alignment of the title and
buttons. Upon doing that, I think we can finalize this and be done with
it.