feat(UI): Implement right click render RC card
I've updated the right click method utilized in Network's NetworkGraph
to render a Material UI dialog containing the RecurserCard object. It
should now appropriately update values based on the user node being
clicked.

There is some warning with child-parent passing of props  noted in the
console but the functionality seems to be in place.