refactor(api): Clean up DB querying
Refactored the get_graph_data() method to minimize SQL queries and clean
up the overall code style. It now works as it did before but with joins,
allow for us to access fields such as location and company names as
opposed to just IDs.

There appears to still be an issue with returning participants but its
functional enough for now.

Could not get the db.py script to work inside the api/ directory but
will add it to the backlog.