feat(UI): Start Recurser search bar implementation
Utilized the Material UI Autocomplete component to start implementing a
Recurser search. The user should be able to have a keyPress of Enter and
have the resulting individual be the focus of the graph.

Working to figure out how to pass props/methods between sibling
components through the Context API.