feat(UI): Polish fuzzy search results dialog
The user should now be able to easily search by criteria and query,
resulting in a list of relevant results (akin to recurse.com/directory).
From here, users can filter by similarity score threshold, number of
results, or simply manually pick nodes as desired.

The visualization still needs to be polished but for now it highlights
selected nodes green and other nodes black. Work with the camera and
link colors is on the to-do list moving forward.