refactor(UI): Split up fuzzy search bar/results
Due to FuzzySearchBar being 400ish lines, I decided it would be best to
split up the actual search bar and the derived results into 2 separate
files.

The idea of passing props from the parent to the child seems a bit
cumbersome here since there is a decent amount of shared state but I'm
going through with it for now.

Perhaps a future refactor is in store but as long as the functionality
is as expected, that task will be put in the backlog.