feat(viz): Implement basic DFS in graph
The graph should now respond to user input and find a target node as
intended. It does not yet work on multiple nodes but should be simple
enough to configure.

Also requires a final path to be displayed on the graph as well as
camera movement along that path when the traversal is over.